spec updated on 2020-08-27 (#28)
* add workaround for #27

* spec updated on 2020-08-27

+++ rail/v3/models/service_d_t_o_version3_rail_t_r_a_line_transfer.go
@@ -41,7 +41,7 @@ type ServiceDTOVersion3RailTRALineTransfer struct {
 
 	// 轉乘耗時參考時間(分)
 	// Required: true
-	MinTransferTime *int32 `json:"MinTransferTime"`
+	MinTransferTime *float64 `json:"MinTransferTime"`
 
 	// 路線間轉乘(迄)之路線代碼
 	// Required: true
