+++ rail/v3/models/service_d_t_o_version3_rail_t_r_a_station_transfer_airport_transfer.go
@@ -39,7 +39,7 @@ type ServiceDTOVersion3RailTRAStationTransferAirportTransfer struct {
 	IsOnSiteTransfer bool `json:"IsOnSiteTransfer,omitempty"`
 
 	// 最短轉乘時間
-	MinTransferTime int32 `json:"MinTransferTime,omitempty"`
+	MinTransferTime float64 `json:"MinTransferTime,omitempty"`
 
 	// 運具種類代碼
 	// Required: true
